// frontend/src/generated-api/model/usersGet200ResponseInner.ts
/**
 * API V1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UsersGet200ResponseInner { 
    id: number;
    first_name: string;
    last_name: string;
    name?: string;
    username: string;
    email: string;
    created_at?: string;
    updated_at?: string;
}